Removed the dependency on Unix's diff command.
